added a init class to the project
The InitSisyphus - Class checks if the needed fonts are installed and
installs them if they don't exist. The fonts are automatically, with
each build, transferred to the running directory in the folder fonts.

the font-configuration files are needed to configure the font for the
mathml generator and have to stay at the root path of our solution!